[#6] chore: 예약하기 버튼 이동
메인 리스트에 있던 예약하기 버튼을 디테일 페이지 내부로 이동하였습니다.
